# public/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/v1/addresses/{address}": {
            "get": {
                "operationId": "addresses.show",
                "tags": [
                    "Address"
                ],
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "required": true,
                        "description": "The address ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`AddressResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AddressResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "put": {
                "operationId": "addresses.update",
                "tags": [
                    "Address"
                ],
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "required": true,
                        "description": "The address ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAddressRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`AddressResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AddressResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "addresses.destroy",
                "tags": [
                    "Address"
                ],
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "required": true,
                        "description": "The address ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/cart/items": {
            "post": {
                "operationId": "cart.addItem",
                "tags": [
                    "Cart"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddToCartRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`CartResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CartResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/v1/cart/items/{product}": {
            "put": {
                "operationId": "cart.updateItem",
                "tags": [
                    "Cart"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCartItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`CartResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CartResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "delete": {
                "operationId": "cart.removeItem",
                "tags": [
                    "Cart"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`CartResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CartResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/checkout": {
            "post": {
                "operationId": "checkout.checkout",
                "tags": [
                    "Checkout"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CheckoutRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "order": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "required": [
                                        "order"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/wishlist/{product}": {
            "post": {
                "operationId": "wishlist.store",
                "tags": [
                    "Wishlist"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "anyOf": [
                                                {
                                                    "type": "null"
                                                },
                                                {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                {
                                                    "type": "string"
                                                }
                                            ]
                                        },
                                        "product_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "product_id"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "delete": {
                "operationId": "wishlist.destroy",
                "tags": [
                    "Wishlist"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "anyOf": [
                                                {
                                                    "type": "null"
                                                },
                                                {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                {
                                                    "type": "string"
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AddToCartRequest": {
                "type": "object",
                "properties": {
                    "product_id": {
                        "type": "integer"
                    },
                    "quantity": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 999
                    }
                },
                "required": [
                    "product_id",
                    "quantity"
                ],
                "title": "AddToCartRequest"
            },
            "AddressResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "label": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "city": {
                        "type": "string"
                    },
                    "address_line": {
                        "type": "string"
                    },
                    "postal_code": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "phone": {
                        "type": "string"
                    },
                    "is_default": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "id",
                    "label",
                    "city",
                    "address_line",
                    "postal_code",
                    "phone",
                    "is_default"
                ],
                "title": "AddressResource"
            },
            "CartItemResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "product_id": {
                        "type": "integer"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "product": {
                        "$ref": "#/components/schemas/ProductResource"
                    },
                    "line_total": {
                        "type": "number"
                    }
                },
                "required": [
                    "id",
                    "product_id",
                    "quantity"
                ],
                "title": "CartItemResource"
            },
            "CartResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CartItemResource"
                        }
                    },
                    "items_count": {
                        "type": "integer"
                    },
                    "subtotal": {
                        "type": "number"
                    }
                },
                "required": [
                    "id",
                    "items_count",
                    "subtotal"
                ],
                "title": "CartResource"
            },
...
            "CheckoutRequest": {
                "type": "object",
                "properties": {
                    "address_id": {
                        "type": "integer"
                    }
                },
                "required": [
                    "address_id"
                ],
                "title": "CheckoutRequest"
            },
...
            "Order": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "address_id": {
                        "type": "integer"
                    },
                    "subtotal": {
                        "type": "string"
                    },
                    "shipping_cost": {
                        "type": "string"
                    },
                    "vat_amount": {
                        "type": "string"
                    },
                    "total_amount": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "payment_status": {
                        "type": "string"
                    },
                    "notes": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "user_id",
                    "address_id",
                    "subtotal",
                    "shipping_cost",
                    "vat_amount",
                    "total_amount",
                    "status",
                    "payment_status",
                    "notes",
                    "created_at",
                    "updated_at"
                ],
                "title": "Order"
            },
            "ProductResource": {
                "type": "object",
                "properties": {
                    "locale": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "category_id": {
                        "type": "integer"
                    },
                    "sku": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "thumbnail_url": {
                        "type": "string"
                    },
                    "category": {
                        "$ref": "#/components/schemas/CategoryResource"
                    },
                    "inventory": {
                        "type": "object",
                        "properties": {
                            "quantity": {
                                "type": "integer"
                            },
                            "low_stock_level": {
                                "type": "integer"
                            },
                            "in_stock": {
                                "type": "boolean"
                            }
                        },
                        "required": [
                            "quantity",
                            "low_stock_level",
                            "in_stock"
                        ]
                    }
                },
                "required": [
                    "locale",
                    "id",
                    "category_id",
                    "sku",
                    "price",
                    "name",
                    "description",
                    "is_active",
                    "thumbnail_url"
                ],
                "title": "ProductResource"
            },
...
            "UpdateAddressRequest": {
                "type": "object",
                "properties": {
                    "label": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 100
                    },
                    "city": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "address_line": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "postal_code": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 20
                    },
                    "phone": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "is_default": {
                        "type": [
                            "boolean",
                            "null"
                        ]
                    }
                },
                "title": "UpdateAddressRequest"
            },
            "UpdateCartItemRequest": {
                "type": "object",
                "properties": {
                    "quantity": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 999
                    }
                },
                "required": [
                    "quantity"
                ],
                "title": "UpdateCartItemRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}